Retrieve Bulk Series Metadata for International Debt Statistics

This function retrieves a tibble with metadata for series available via bulk download of the World Bank International Debt Statistics (IDS).

Usage

ids_bulk_series()

Value

A tibble containing the available series and their metadata:

series_id

The unique identifier for the series (e.g., "BN.CAB.XOKA.CD").

series_name

The name of the series (e.g., "Current account balance (current US$)").

source_id

The ID of the data source providing the indicator.

source_name

The name or description of the source of the indicator data.

source_note

Additional notes or descriptions about the data source.

source_organization

The organization responsible for the data source.

Examples

ids_bulk_series()
#> # A tibble: 495 × 6
#>    series_id   series_name source_id source_name source_note source_organization
#>    <chr>       <chr>           <int> <chr>       <chr>       <chr>              
#>  1 DT.UND.DPP… Undisburse…        81 " Internat… Undisburse… World Bank, Intern…
#>  2 DT.UND.OFF… Undisburse…        81 " Internat… Undisburse… World Bank, Intern…
#>  3 NY.GNP.MKT… GNI (curre…         2 "World Dev… GNI (forme… World Bank nationa…
#>  4 DT.INT.BLA… PS, bilate…        81 " Internat… Public sec… World Bank, Intern…
#>  5 DT.TDS.MLT… GG, multil…        81 " Internat… General go… World Bank, Intern…
#>  6 DT.TDS.MLT… OPS, multi…        81 " Internat… Other publ… World Bank, Intern…
#>  7 DT.TDS.OFF… PPG, offic…        81 " Internat… Public and… World Bank, Intern…
#>  8 DT.TDS.OFF… PRVG, offi…        81 " Internat… Private se… World Bank, Intern…
#>  9 DT.TDS.OFF… PS, offici…        81 " Internat… Public sec… World Bank, Intern…
#> 10 DT.TDS.PBN… OPS, bonds…        81 " Internat… Other publ… World Bank, Intern…
#> # ℹ 485 more rows