List Available Geographies from the World Bank International Debt Statistics API

This function returns a tibble with available geographies from the World Bank International Debt Statistics (IDS) API. Each row provides details on geographies, including their unique identifiers, names, and types.

Usage

ids_list_geographies()

Value

A tibble containing the available geographies and their attributes:

geography_id

ISO 3166-1 alpha-3 code of the geography (e.g., "ZMB").

geography_name

The standardized name of the geography (e.g., "Zambia").

geography_iso2code

ISO 3166-1 alpha-2 code of the geography (e.g., "ZM").

geography_type

The type of geography (e.g., "Country", "Region").

capital_city

The capital city of the geography (e.g., "Lusaka").

region_id

The unique identifier for the region (e.g., "SSF").

region_iso2code

ISO 3166-1 alpha-2 code of the region (e.g., "ZG").

region_name

The name of the region (e.g., "Sub-Saharan Africa").

admin_region_id

The unique identifier for the administrative region (e.g., "SSA").

admin_region_iso2code

The ISO 3166-1 alpha-2 code for the administrative region (e.g., "ZF").

admin_region_name

The name of the administrative region (e.g., "Sub-Saharan Africa (excluding high income)").

lending_type_id

The unique identifier for the lending type (e.g., "IDX").

lending_type_iso2code

ISO code for the lending type (e.g., "XI").

lending_type_name

The name of the lending type (e.g., "IDA").

Examples

ids_list_geographies()
#> # A tibble: 136 × 17
#>    geography_id geography_iso2code geography_type capital_city geography_name
#>    <chr>        <chr>              <chr>          <chr>        <chr>         
#>  1 AFG          AF                 Country        Kabul        Afghanistan   
#>  2 AGO          AO                 Country        Luanda       Angola        
#>  3 ALB          AL                 Country        Tirane       Albania       
#>  4 ARG          AR                 Country        Buenos Aires Argentina     
#>  5 ARM          AM                 Country        Yerevan      Armenia       
#>  6 AZE          AZ                 Country        Baku         Azerbaijan    
#>  7 BDI          BI                 Country        Bujumbura    Burundi       
#>  8 BEN          BJ                 Country        Porto-Novo   Benin         
#>  9 BFA          BF                 Country        Ouagadougou  Burkina Faso  
#> 10 BGD          BD                 Country        Dhaka        Bangladesh    
#> # ℹ 126 more rows
#> # ℹ 12 more variables: region_id <chr>, region_iso2code <chr>,
#> #   region_name <chr>, admin_region_id <chr>, admin_region_iso2code <chr>,
#> #   admin_region_name <chr>, income_level_id <chr>,
#> #   income_level_iso2code <chr>, income_level_name <chr>,
#> #   lending_type_id <chr>, lending_type_iso2code <chr>, lending_type_name <chr>